requirements.txt
streamlit>=1.37.0
google-generativeai>=0.8.5
gTTS>=2.5.4
PyPDF2>=3.0.0
pandas>=2.2.0 
